+++ packages/vendure-plugin-invoices/src/api/invoice-admin.resolver.ts
@@ -5,13 +5,18 @@ import {
   Ctx,
   ID,
   OrderService,
+  PaginatedList,
   RequestContext,
   UserInputError,
 } from '@vendure/core';
 import { PLUGIN_INIT_OPTIONS } from '../constants';
 import { InvoicePluginConfig } from '../index';
 import { InvoiceService } from '../services/invoice.service';
-import { Invoice, InvoiceConfigInput } from '../ui/generated/graphql';
+import {
+  Invoice,
+  InvoiceConfigInput,
+  QueryInvoicesArgs,
+} from '../ui/generated/graphql';
 import { InvoiceConfigEntity } from '../entities/invoice-config.entity';
 import { invoicePermission } from './invoice-common.resolver';
 
@@ -44,13 +49,16 @@ export class InvoiceAdminResolver {
         `Can not generate invoice for an order without 'customer.emailAddress'`
       );
     }
-    const invoice = await this.invoiceService.createAndSaveInvoice(
+    const invoice = await this.invoiceService.createInvoicesForOrder(
       ctx.channel.token,
-      order.code
+      order.code,
+      false
     );
     return {
       ...invoice,
       isCreditInvoice: invoice.isCreditInvoice,
+      orderId: order.id,
+      orderCode: order.code,
       downloadUrl: this.invoiceService.getDownloadUrl(
         ctx,
         invoice,
@@ -66,5 +74,14 @@ export class InvoiceAdminResolver {
     @Ctx() ctx: RequestContext
   ): Promise<InvoiceConfigEntity | undefined> {
     return this.invoiceService.getConfig(ctx);
+  }
+
+  @Query()
+  @Allow(invoicePermission.Permission)
+  async invoices(
+    @Ctx() ctx: RequestContext,
+    @Args() args: QueryInvoicesArgs
+  ): Promise<PaginatedList<Invoice>> {
+    return this.invoiceService.findAll(ctx, args.options || undefined);
   }
 }