+++ packages/vendure-scripts/src/index.ts
@@ -1,1 +1,1 @@
-export * from './assign-all-products-to-channel';
+export * from './assign-products-to-channel';