+++ packages/vendure-plugin-invoices/CHANGELOG.md
@@ -1,3 +1,19 @@
+# 2.3.1 (2024-06-05)
+
+- Sort invoices by created at
+
+# 2.3.0 (2024-06-05)
+
+- Invoice listing page and bulk download of invoices
+
+# 2.2.0 (2024-06-04)
+
+- When an order is cancelled, generate credit invoice, and no invoice regeneration after that
+
+# 2.1.1 (2024-06-04)
+
+- Prevent the possibility of connecting the wrong invoice file to an order with concurrent invoice generation. See #433
+
 # 2.1.0 (2024-02-14)
 
 - Work on ui improvements as described in #352